pyproject.toml
[project]
name = "eidos-framework"
version = "1.0.0"
description = "The Physics of Software"
authors = [
    {name = "HENG Yang", email = "cosmos@eidos.null"},
]
dependencies = [
    "numpy",
    "polars>=1.0.0",
    "pyarrow>=16.0.0",
    "structlog>=24.1.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "beartype>=0.16.0",
]
requires-python = ">=3.14"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/HengYangDS/eidos-framework"
Repository = "https://github.com/HengYangDS/eidos-framework"

[project.optional-dependencies]
# Distributed Computing
ray = [
    "ray[default]>=2.0.0",
]

# Financial Mathematics (uses core Polars/Numpy)
quant = []

# Install only when you need to expose pipelines as HTTP APIs
http = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
]

# Observability and telemetry (optional)
otlp = [
    "opentelemetry-exporter-otlp>=1.27",
]

# CSV sink relies on Python stdlib; this extra reserved for future connectors
csv = []
sql = [
    "pymssql>=2.2",
]
io = [
    "kafka-python>=2.0.0",
    "redis>=5.0.0",
]
dev = [
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
    "syrupy>=4.6.0",
]

[project.scripts]
eidos = "eidos.dx.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
